// src/data/ListingData.d.ts
export declare const content: {
    title: string;
    link: string;
    video: string;
    image: string;
    alt: string;
    date: string;
    startDate: string;
    endDate: string;
    excerpt: string;
    tags: {
        category: {
            id: number;
            name: string;
            slug: string;
        }[];
    };
};